requirements.txt
allure-pytest==2.15.0
allure-python-commons==2.15.0
attrs==25.4.0
certifi==2025.10.5
cffi==2.0.0
colorama==0.4.6
Faker==37.11.0
h11==0.16.0
idna==3.11
iniconfig==2.3.0
outcome==1.3.0.post0
packaging==25.0
pluggy==1.6.0
pycparser==2.23
Pygments==2.19.2
PySocks==1.7.1
pytest==8.4.2
python-dotenv==1.1.1
selenium==4.37.0
sniffio==1.3.1
sortedcontainers==2.4.0
trio==0.31.0
trio-websocket==0.12.2
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
websocket-client==1.9.0
wsproto==1.2.0
